refactor(ai-service): decouple from telegram dependency
ai-service changes:
- Remove bot/ directory (handler.py depended on telegram)
- Update __init__.py to export core modules only
- Now exports: run_analysis, PromptRegistry, build_prompt, call_llm, fetch_payload

telegram-service changes:
- Rewrite bot/ai_integration.py with Telegram UI logic
- Import ai-service core via sys.path
- AIAnalysisHandler now lives in telegram-service

Architecture:
- ai-service: pure AI analysis logic (no telegram dependency)
- telegram-service: Telegram UI + calls ai-service via adapter
