+++ services/ai-service/src/__init__.py
@@ -1,21 +1,22 @@
 # -*- coding: utf-8 -*-
 """
-AI 分析服务
+AI 分析服务（独立服务）
 
-作为 telegram-service 的子模块集成，提供 AI 深度分析功能。
+核心功能：
+- run_analysis: 执行 AI 分析
+- PromptRegistry: 提示词注册表
+- call_llm: 调用 LLM
+- fetch_payload: 获取分析数据
 """
-from src.bot import (
-    AIAnalysisHandler,
-    get_ai_handler,
-    register_ai_handlers,
-)
 from src.pipeline import run_analysis
-from src.prompt import PromptRegistry
+from src.prompt import PromptRegistry, build_prompt
+from src.llm import call_llm
+from src.data import fetch_payload
 
 __all__ = [
-    "AIAnalysisHandler",
-    "get_ai_handler",
-    "register_ai_handlers",
     "run_analysis",
     "PromptRegistry",
+    "build_prompt",
+    "call_llm",
+    "fetch_payload",
 ]
