--- services/ai-service/src/bot/__init__.py
@@ -1,17 +1,0 @@
-# -*- coding: utf-8 -*-
-"""AI 分析 Bot 模块"""
-from .handler import (
-    AIAnalysisHandler,
-    get_ai_handler,
-    register_ai_handlers,
-    SELECTING_COIN,
-    SELECTING_INTERVAL,
-)
-
-__all__ = [
-    "AIAnalysisHandler",
-    "get_ai_handler",
-    "register_ai_handlers",
-    "SELECTING_COIN",
-    "SELECTING_INTERVAL",
-]
